@@ -22,7 +22,7 @@
   </modules>
 
   <properties>
-    <maven.compiler.release>11</maven.compiler.release>
+    <maven.compiler.release>17</maven.compiler.release>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <docs.directory>${project.basedir}/docs</docs.directory>
@@ -34,7 +34,7 @@
         <plugin>
           <groupId>org.codehaus.mojo</groupId>
           <artifactId>license-maven-plugin</artifactId>
-          <version>2.2.0</version>
+          <version>2.5.0</version>
           <configuration>
             <cleanLicensesOutputDirectory>true</cleanLicensesOutputDirectory>
             <errorRemedy>ignore</errorRemedy>
@@ -47,8 +47,42 @@
             <sortByGroupIdAndArtifactId>true</sortByGroupIdAndArtifactId>
           </configuration>
         </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-enforcer-plugin</artifactId>
+          <version>3.5.0</version>
+          <configuration>
+            <rules>
+              <requireMavenVersion>
+                <!--
+                  Align with requirements from Blueprint workspace.
+                  Possibly validate via `mvn versions:display-plugin-updates`
+                  regarding required minimum versions.
+                -->
+                <version>3.8.6</version>
+              </requireMavenVersion>
+              <requireJavaVersion>
+                <version>[17,)</version>
+              </requireJavaVersion>
+            </rules>
+          </configuration>
+        </plugin>
       </plugins>
     </pluginManagement>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-enforcer-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>enforce-versions</id>
+            <goals>
+              <goal>enforce</goal>
+            </goals>
+          </execution>
+        </executions>
+      </plugin>
+    </plugins>
   </build>
 
   <profiles>
