@@ -4,8 +4,10 @@ import com.coremedia.labs.translation.gcc.facade.GCExchangeFacade;
 import com.coremedia.labs.translation.gcc.facade.GCExchangeFacadeSessionProvider;
 import com.coremedia.labs.translation.gcc.facade.GCFacadeCommunicationException;
 import com.coremedia.labs.translation.gcc.facade.GCSubmissionModel;
+import com.coremedia.labs.translation.gcc.facade.GCSubmissionState;
 import com.coremedia.labs.translation.gcc.facade.GCTaskModel;
-import com.google.common.annotations.VisibleForTesting;
+import com.coremedia.labs.translation.gcc.facade.mock.settings.MockError;
+import com.coremedia.labs.translation.gcc.facade.mock.settings.MockSettings;
 import edu.umd.cs.findbugs.annotations.DefaultAnnotation;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
@@ -19,7 +21,6 @@ import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.time.ZonedDateTime;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -32,47 +33,36 @@ import static java.util.stream.Collectors.toList;
 import static org.slf4j.LoggerFactory.getLogger;
 
 /**
- * <p>
  * This facade will mock the behavior of GCC. It is especially meant for
  * demo cases and for testing purpose.
- * </p>
  * <p>
  * To get an instance of this facade, use {@link GCExchangeFacadeSessionProvider}.
- * </p>
  */
 @DefaultAnnotation(NonNull.class)
 public final class MockedGCExchangeFacade implements GCExchangeFacade {
   private static final Logger LOG = getLogger(lookup().lookupClass());
 
   private static final ContentStore contentStore = new ContentStore();
-  private static final SubmissionStore submissionStore = new SubmissionStore();
-  @Nullable
-  private MockGCExchangeFacadeProvider.MockError mockError = null;
+  /**
+   * The submission store to keep track of all submissions.
+   * <p>
+   * This must be a singleton instance, as it is used to keep track of all
+   * submissions across multiple facade instances.
+   */
+  private static final SubmissionStore submissionStore = SubmissionStore.getInstance();
+  @NonNull
+  private final MockSettings mockSettings;
 
-  MockedGCExchangeFacade() {
-  }
-
-  @SuppressWarnings("SameParameterValue")
-  @VisibleForTesting
-  MockedGCExchangeFacade setDelayBaseSeconds(long delayBaseSeconds) {
-    submissionStore.setDelayBaseSeconds(delayBaseSeconds);
-    return this;
-  }
-
-  @SuppressWarnings({"SameParameterValue", "UnusedReturnValue"})
-  @VisibleForTesting
-  MockedGCExchangeFacade setDelayOffsetPercentage(int delayOffsetPercentage) {
-    submissionStore.setDelayOffsetPercentage(delayOffsetPercentage);
-    return this;
-  }
-
-  void setMockError(@Nullable MockGCExchangeFacadeProvider.MockError mockError) {
-    this.mockError = mockError;
+  MockedGCExchangeFacade(@NonNull MockSettings mockSettings) {
+    this.mockSettings = mockSettings;
+    // By intention may adapt the settings also within the submission store
+    // on each new instance creation of the facade. While this is not meant
+    // to update the settings for existing/running submissions and tasks,
+    // any update will be applied to new submissions and tasks.
+    submissionStore.applySettings(mockSettings);
   }
 
   /**
-   * {@inheritDoc}
-   *
    * @throws UnsupportedOperationException if called
    * @implNote Will throw an {@link UnsupportedOperationException} as there is no delegate available.
    */
@@ -83,7 +73,7 @@ public final class MockedGCExchangeFacade implements GCExchangeFacade {
 
   @Override
   public String uploadContent(String fileName, Resource resource, Locale sourceLocale) {
-    if (mockError == MockGCExchangeFacadeProvider.MockError.UPLOAD_COMMUNICATION) {
+    if (mockSettings.error() == MockError.UPLOAD_COMMUNICATION) {
       throw new GCFacadeCommunicationException("Exception to test upload communication errors with translation service.");
     }
     return contentStore.addContent(resource);
@@ -93,23 +83,23 @@ public final class MockedGCExchangeFacade implements GCExchangeFacade {
   public long submitSubmission(@Nullable String subject, @Nullable String comment, ZonedDateTime dueDate, @Nullable String workflow, @Nullable String submitter, Locale sourceLocale, Map<String, List<Locale>> contentMap) {
     String trimmedSubject = Objects.toString(subject, "").trim();
     List<SubmissionContent> collect = contentMap.entrySet().stream()
-            .map(e -> new SubmissionContent(
-                    e.getKey(),
-                    // Reads and removes the content.
-                    contentStore.removeContent(e.getKey()),
-                    e.getValue()))
-            .collect(toList());
+      .map(e -> new SubmissionContent(
+        e.getKey(),
+        // Reads and removes the content.
+        contentStore.removeContent(e.getKey()),
+        e.getValue()))
+      .collect(toList());
     return submissionStore.addSubmission(trimmedSubject, collect);
   }
 
   @Override
   public int cancelSubmission(long submissionId) {
-    if (mockError == MockGCExchangeFacadeProvider.MockError.CANCEL_COMMUNICATION) {
+    if (mockSettings.error() == MockError.CANCEL_COMMUNICATION) {
       throw new GCFacadeCommunicationException("Exception to test cancel communication errors with translation service.");
     }
-    if (mockError == MockGCExchangeFacadeProvider.MockError.CANCEL_RESULT) {
+    if (mockSettings.error() == MockError.CANCEL_RESULT) {
       // Any one of the possible errors documented in
-      // https://connect-dev.translations.com/docs/api/GlobalLink_Connect_Cloud_API_Documentation.htm#submissions_cancel
+      // https://connect-dev.translations.com/docs/api/v2/index.html#submissions_cancel
       // 400, 401, 404, 500
       return 404;
     }
@@ -133,12 +123,12 @@ public final class MockedGCExchangeFacade implements GCExchangeFacade {
   }
 
   private void downloadTask(Task task, BiPredicate<? super InputStream, ? super GCTaskModel> taskDataConsumer) {
-    if (mockError == MockGCExchangeFacadeProvider.MockError.DOWNLOAD_COMMUNICATION) {
+    if (mockSettings.error() == MockError.DOWNLOAD_COMMUNICATION) {
       throw new GCFacadeCommunicationException("Exception to test download communication errors with translation service.");
     }
     boolean success = false;
     String untranslatedContent = task.getContent();
-    String translatedContent = TranslationUtil.translateXliff(untranslatedContent, mockError == MockGCExchangeFacadeProvider.MockError.DOWNLOAD_XLIFF);
+    String translatedContent = TranslationUtil.translateXliff(untranslatedContent, mockSettings.error() == MockError.DOWNLOAD_XLIFF);
     try (InputStream is = new ByteArrayInputStream(translatedContent.getBytes(StandardCharsets.UTF_8))) {
       success = taskDataConsumer.test(is, new GCTaskModel(task.getId(), task.getTargetLocale()));
     } catch (IOException e) {
@@ -155,9 +145,15 @@ public final class MockedGCExchangeFacade implements GCExchangeFacade {
     cancelledTasks.forEach(Task::markAsCancellationConfirmed);
   }
 
+
   @Override
   public GCSubmissionModel getSubmission(long submissionId) {
-    return new GCSubmissionModel(submissionId, Collections.singletonList(Long.toString(submissionId)), submissionStore.getSubmissionState(submissionId));
+    // State: Will query an exception if not found.
+    GCSubmissionState submissionState = submissionStore.getSubmissionState(submissionId);
+    return GCSubmissionModel.builder(submissionId)
+      .pdSubmissionIds(List.of(Long.toString(submissionId)))
+      .state(submissionState)
+      .error(mockSettings.error() == MockError.SUBMISSION_ERROR)
+      .build();
   }
-
 }
