@@ -14,16 +14,18 @@ ResourceBundleUtil.override(GccProcessDefinitions_properties, {
   TranslationGlobalLink_submission_status_IN_PROGRESS: "In Bearbeitung",
   TranslationGlobalLink_submission_status_CANCELLED: "Abgebrochen",
   TranslationGlobalLink_submission_status_CANCELLATION_CONFIRMED: "Abbruch bestätigt",
+  TranslationGlobalLink_submission_status_CANCELLATION_REQUESTED: "Abbruch beantragt",
   TranslationGlobalLink_submission_status_TRANSLATE: "Übersetzung",
   TranslationGlobalLink_submission_status_REVIEW: "Überprüfung",
   TranslationGlobalLink_submission_status_COMPLETED: "Abgeschlossen",
+  TranslationGlobalLink_submission_status_REDELIVERED: "Erneut ausgeliefert",
   TranslationGlobalLink_submission_status_DELIVERED: "Ausgeliefert",
   TranslationGlobalLink_submission_status_OTHER: "Unbekannt",
   TranslationGlobalLink_submission_status_unavailable: "Nicht verfügbar",
   TranslationGlobalLink_submission_dueDate_tooltip: "Bitte Datum in der Zukunft auswählen.",
   translationResultXliff_Button_text: "ZIP Datei mit Detailinformationen herunterladen",
-  confirm_cancellation_title: "Üersetzungsaufträge Abbrechen",
-  confirm_cancellation: "Die ausgewählten Üersetzungsaufträge werden abgebrochen. Wollen Sie fortfahren?",
+  confirm_cancellation_title: "Übersetzungsaufträge Abbrechen",
+  confirm_cancellation: "Die ausgewählten Übersetzungsaufträge werden abgebrochen. Wollen Sie fortfahren?",
   Action_Tooltip_Cancel_Process: "Die ausgewählten Übersetzungsaufträge abbrechen.",
   TranslationGlobalLink_completed_Locales: "Bereits übersetzte Locales",
   TranslationGlobalLink_Multi_Target_Locale_Text: "{0} Locales",
