@@ -18,6 +18,7 @@ import static java.time.ZoneOffset.UTC;
  */
 @DefaultAnnotation(NonNull.class)
 final class GCUtil {
+
   private GCUtil() {
     // Utility class
   }
@@ -29,6 +30,7 @@ final class GCUtil {
    * @param dateTime date-time object to convert
    * @return date in UTC timezone
    */
+  @SuppressWarnings("UseOfObsoleteDateTimeApi")
   static Date toUnixDateUtc(ZonedDateTime dateTime) {
     ZonedDateTime utcDateTime = dateTime.withZoneSameInstant(UTC);
     return Date.from(utcDateTime.toInstant());
@@ -36,19 +38,19 @@ final class GCUtil {
 
   /**
    * Executes the {@link PageableRequest} until all results were received. Note,
-   * that because of asynchronous updates on server this method cannot guarantee
-   * to process all pages as well as some results may be duplicated because
-   * the moved to a different page meanwhile.
+   * that because of asynchronous updates on the server this method cannot
+   * guarantee to process all pages as well as some results may be duplicated
+   * because the moved to a different page meanwhile.
    *
    * @param rawRequestSupplier creates the raw request with default paging configuration; note, that you may override
    *                           the page size for this raw request which will be taken into account during pagination
    * @param requestExecutor    executes the requests; RuntimeExceptions thrown by this executor will be wrapped into a
    *                           {@link GCFacadeCommunicationException}; the returned response is just used to extract
-   *                           the total page number, so it is task of the executor to merge all results
+   *                           the total page number, so it is the task of the executor to merge all results
    * @param <I>                the request type
    * @param <O>                the response type
    * @throws GCFacadeCommunicationException if pagination has been interrupted by a {@code RuntimeException} during
-   *                           request processing
+   *                                        request processing
    */
   static <I extends PageableRequest, O extends PageableResponseData> void processAllPages(Supplier<I> rawRequestSupplier, Function<I, O> requestExecutor) {
     // Initial Page Number
