@@ -13,7 +13,6 @@ import org.springframework.core.io.Resource;
 
 import java.io.InputStream;
 import java.time.ZonedDateTime;
-import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -43,8 +42,6 @@ public final class DisabledGCExchangeFacade implements GCExchangeFacade {
   }
 
   /**
-   * {@inheritDoc}
-   *
    * @throws UnsupportedOperationException if called
    * @implNote Will throw an {@link UnsupportedOperationException} as there is no delegate available.
    */
@@ -85,7 +82,7 @@ public final class DisabledGCExchangeFacade implements GCExchangeFacade {
 
   @Override
   public GCSubmissionModel getSubmission(long submissionId) {
-    return new GCSubmissionModel(submissionId, Collections.emptyList());
+    return GCSubmissionModel.builder(submissionId).build();
   }
 
   private static GCFacadeException createDisabledException() {
