@@ -5,7 +5,9 @@ import com.coremedia.labs.translation.gcc.facade.GCConfigProperty;
 import com.coremedia.labs.translation.gcc.facade.GCExchangeFacade;
 import com.coremedia.labs.translation.gcc.facade.GCSubmissionState;
 import com.coremedia.labs.translation.gcc.facade.GCTaskModel;
+import com.coremedia.labs.translation.gcc.facade.mock.settings.MockSettings;
 import com.google.common.io.ByteSource;
+import edu.umd.cs.findbugs.annotations.NonNull;
 import org.assertj.core.api.Assertions;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
@@ -21,6 +23,7 @@ import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Locale;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.function.BiPredicate;
 
@@ -32,7 +35,7 @@ import static org.slf4j.LoggerFactory.getLogger;
 
 class MockedGCExchangeFacadeTest {
   private static final Logger LOG = getLogger(lookup().lookupClass());
-  private static final int TRANSLATION_TIMEOUT_MINUTES = 3;
+  private static final long TRANSLATION_TIMEOUT_MINUTES = 3L;
 
   private static final String XLIFF_FILE="LoremIpsum.xliff";
 
@@ -45,15 +48,19 @@ class MockedGCExchangeFacadeTest {
 
     Resource xliffResource = new ClassPathResource(XLIFF_FILE, MockedGCExchangeFacade.class);
 
-    MockedGCExchangeFacade facade = new MockedGCExchangeFacade();
     // Let the tasks proceed faster.
-    facade.setDelayBaseSeconds(2).setDelayOffsetPercentage(20);
+    GCExchangeFacade facade = new MockedGCExchangeFacade(MockSettings.fromMockConfig(
+      Map.of(
+        MockSettings.CONFIG_STATE_CHANGE_DELAY_SECONDS, 2L,
+        MockSettings.CONFIG_STATE_CHANGE_DELAY_OFFSET_PERCENTAGE, 20
+      )
+    ));
 
     String fileId = facade.uploadContent(testName, xliffResource, null);
     long submissionId = facade.submitSubmission(
             testName,
             null,
-            ZonedDateTime.of(LocalDateTime.now().plusHours(2), ZoneId.systemDefault()),
+            ZonedDateTime.of(LocalDateTime.now().plusHours(2L), ZoneId.systemDefault()),
             null,
             "admin",
             Locale.US, singletonMap(fileId, singletonList(Locale.ROOT)));
@@ -74,17 +81,12 @@ class MockedGCExchangeFacadeTest {
 
   }
 
-  private class TaskDataConsumer implements BiPredicate<InputStream, GCTaskModel> {
-    private final StringBuilder xliffResult;
-
-    TaskDataConsumer(StringBuilder xliffResult) {
-      this.xliffResult = xliffResult;
-    }
-
+  private record TaskDataConsumer(StringBuilder xliffResult) implements BiPredicate<InputStream, GCTaskModel> {
     @Override
     public boolean test(InputStream is, GCTaskModel task) {
       ByteSource byteSource = new ByteSource() {
         @Override
+        @NonNull
         public InputStream openStream() {
           return is;
         }
@@ -104,15 +106,19 @@ class MockedGCExchangeFacadeTest {
 
     Resource xliffResource = new ClassPathResource(XLIFF_FILE, MockedGCExchangeFacade.class);
 
-    MockedGCExchangeFacade facade = new MockedGCExchangeFacade();
     // Let the tasks proceed faster.
-    facade.setDelayBaseSeconds(2).setDelayOffsetPercentage(20);
+    GCExchangeFacade facade = new MockedGCExchangeFacade(MockSettings.fromMockConfig(
+      Map.of(
+        MockSettings.CONFIG_STATE_CHANGE_DELAY_SECONDS, 2L,
+        MockSettings.CONFIG_STATE_CHANGE_DELAY_OFFSET_PERCENTAGE, 20
+      )
+    ));
 
     String fileId = facade.uploadContent(testName, xliffResource, Locale.US);
     long submissionId = facade.submitSubmission(
             "states:other,cancelled",
             null,
-            ZonedDateTime.of(LocalDateTime.now().plusHours(2), ZoneId.systemDefault()),
+            ZonedDateTime.of(LocalDateTime.now().plusHours(2L), ZoneId.systemDefault()),
             null,
             "admin",
             Locale.US, singletonMap(fileId, singletonList(Locale.ROOT)));
@@ -129,8 +135,8 @@ class MockedGCExchangeFacadeTest {
   private static void assertSubmissionReachesState(GCExchangeFacade facade, long submissionId, GCSubmissionState desiredState) {
     Awaitility.await("Wait for translation to reach state: " + desiredState)
       .atMost(TRANSLATION_TIMEOUT_MINUTES, TimeUnit.MINUTES)
-      .pollDelay(1, TimeUnit.SECONDS)
-      .pollInterval(1, TimeUnit.SECONDS)
+      .pollDelay(1L, TimeUnit.SECONDS)
+      .pollInterval(1L, TimeUnit.SECONDS)
       .conditionEvaluationListener(condition -> LOG.info("Submission {}, Current State: {}, elapsed time in seconds: {}", submissionId, facade.getSubmission(submissionId).getState(), condition.getElapsedTimeInMS() / 1000L))
       .untilAsserted(() -> assertThat(facade.getSubmission(submissionId).getState()).isEqualTo(desiredState));
   }
