@@ -15,12 +15,12 @@ import java.util.Set;
 import java.util.function.BiPredicate;
 
 /**
- * <p>
  * Facade for {@link org.gs4tr.gcc.restclient.GCExchange}.
- * </p>
  * <p>
  * Get instances of facades via a {@link GCExchangeFacadeSessionProvider}.
- * </p>
+ *
+ * @implSpec Implementations of this interface must be stateless, as they are
+ * recreated several times during the lifecycle of the translation workflow.
  */
 @DefaultAnnotation(NonNull.class)
 public interface GCExchangeFacade {
@@ -37,6 +37,9 @@ public interface GCExchangeFacade {
   /**
    * Uploads the given content, like for example an XLIFF file. Returns
    * the file ID which is later required in {@link #submitSubmission(String, String, ZonedDateTime, String, String, Locale, Map)}.
+   * <p>
+   * Note, that {@code fileName} is only allowed to contain characters from
+   * the Unicode Basic Multilingual Plane (BMP).
    *
    * @param fileName the filename of the resource
    * @param resource the resource to send
@@ -144,7 +147,7 @@ public interface GCExchangeFacade {
    *
    * @param submissionId ID of the submission
    * @return the submission model
-   * @throws GCFacadeCommunicationException if the submission could not be retrieved
+   * @throws GCFacadeSubmissionNotFoundException if the submission could not be retrieved
    */
   GCSubmissionModel getSubmission(long submissionId);
 }
