@@ -1,9 +1,7 @@
 package com.coremedia.labs.translation.gcc.workflow;
 
 import com.coremedia.cap.common.CapConnection;
-import com.coremedia.cap.common.CapSession;
 import com.coremedia.cap.common.IdHelper;
-import com.coremedia.cap.common.pool.CapSessionPool;
 import com.coremedia.cap.content.Content;
 import com.coremedia.cap.content.ContentType;
 import com.coremedia.cap.content.Version;
@@ -16,13 +14,13 @@ import com.coremedia.labs.translation.gcc.facade.GCTaskModel;
 import com.coremedia.translate.workflow.AsRobotUser;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Resources;
+import edu.umd.cs.findbugs.annotations.NonNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInfo;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.stubbing.Answer;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,14 +36,17 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.Serial;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Locale;
+import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiPredicate;
-import java.util.function.Supplier;
 import java.util.regex.Pattern;
 
 import static com.coremedia.cap.translate.xliff.XliffImportIssueSeverity.MAJOR;
@@ -60,13 +61,16 @@ import static java.util.Objects.requireNonNull;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.Mockito.doAnswer;
+import static org.mockito.Mockito.doCallRealMethod;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
 import static org.springframework.beans.factory.config.BeanDefinition.SCOPE_SINGLETON;
 import static org.springframework.test.annotation.DirtiesContext.ClassMode.AFTER_CLASS;
 
 /**
  * Tests {@link DownloadFromGlobalLinkAction}.
  */
-@SuppressWarnings("UnstableApiUsage")
 @ExtendWith(SpringExtension.class)
 @ExtendWith(MockitoExtension.class)
 @ContextConfiguration(classes = DownloadFromGlobalLinkActionTest.LocalConfig.class)
@@ -107,6 +111,26 @@ class DownloadFromGlobalLinkActionTest {
   }
 
   @Nested
+  class SubmissionErrorHandling {
+    @Test
+    void shouldForwardSubmissionErrorAsIssue(@Autowired DownloadFromGlobalLinkAction action,
+                                             @Autowired GCExchangeFacade gcExchangeFacade) {
+      doReturn(
+          GCSubmissionModel.builder(1L)
+            .state(GCSubmissionState.IN_PRE_PROCESS)
+            .error(true)
+            .build())
+        .when(gcExchangeFacade).getSubmission(anyLong());
+
+      AtomicReference<DownloadFromGlobalLinkAction.Result> resultHolder = new AtomicReference<>();
+      Map<String, List<Content>> issues = new HashMap<>();
+      action.doExecuteGlobalLinkAction(new DownloadFromGlobalLinkAction.Parameters(1L, new HashSet<>(), false), resultHolder::set, gcExchangeFacade, issues);
+
+      assertThat(issues).containsKey(GlobalLinkWorkflowErrorCodes.SUBMISSION_ERROR);
+    }
+  }
+
+  @Nested
   @DisplayName("Tests error handling on XLIFF import.")
   @ContextConfiguration(classes = DownloadFromGlobalLinkActionTest.LocalConfig.class)
   @DirtiesContext(classMode = AFTER_CLASS)
@@ -125,9 +149,9 @@ class DownloadFromGlobalLinkActionTest {
 
       assertThat(result.resultItems.values().stream().anyMatch(list -> {
         XliffImportResultItem resultItem = list.isEmpty() ? null : list.get(0);
-        return resultItem != null && resultItem.getCode().equals(NO_SUCH_PROPERTY) &&
-                resultItem.getSeverity().equals(MAJOR) && resultItem.getContent().equals(targetContent) &&
-                resultItem.getProperty().equals("nosuchproperty");
+        return resultItem != null && resultItem.getCode() == NO_SUCH_PROPERTY &&
+          resultItem.getSeverity() == MAJOR && Objects.equals(resultItem.getContent(), targetContent) &&
+          Objects.equals(resultItem.getProperty(), "nosuchproperty");
       })).isTrue();
       assertThat(result.resultItems).containsKey(1L);
       assertThat(targetContent.getString("string")).isEmpty();
@@ -135,13 +159,13 @@ class DownloadFromGlobalLinkActionTest {
   }
 
   private static void mockXliffDownload(@Autowired GCExchangeFacade gcExchangeFacade, String xliff) {
-    Mockito.doAnswer((Answer<Boolean>) invocationOnMock -> {
+    doAnswer((Answer<Boolean>) invocationOnMock -> {
       BiPredicate<InputStream, GCTaskModel> consumer = invocationOnMock.getArgument(1);
       ByteArrayInputStream inputStream = new ByteArrayInputStream(xliff.getBytes(StandardCharsets.UTF_8));
       consumer.test(inputStream, new GCTaskModel(1L, Locale.GERMANY));
       return true;
     }).when(gcExchangeFacade).downloadCompletedTasks(anyLong(), any());
-    Mockito.doReturn(new GCSubmissionModel(1, Collections.emptyList(), GCSubmissionState.DELIVERED)).when(gcExchangeFacade).getSubmission(anyLong());
+    doReturn(GCSubmissionModel.builder(1L).state(GCSubmissionState.DELIVERED).build()).when(gcExchangeFacade).getSubmission(anyLong());
   }
 
   private static String readXliff(Version masterVersion, Content targetContent) throws IOException {
@@ -166,6 +190,7 @@ class DownloadFromGlobalLinkActionTest {
   }
 
   private static final class MockedDownloadFromGlobalLinkAction extends DownloadFromGlobalLinkAction {
+    @Serial
     private static final long serialVersionUID = -4082795575498550151L;
     private final ApplicationContext applicationContext;
 
@@ -174,14 +199,15 @@ class DownloadFromGlobalLinkActionTest {
     }
 
     @Override
+    @NonNull
     protected ApplicationContext getSpringContext() {
       return applicationContext;
     }
 
     @Override
     AsRobotUser getAsRobotUser() {
-      AsRobotUser asRobotUser = Mockito.mock(AsRobotUser.class);
-      Mockito.doCallRealMethod().when(asRobotUser).call(any());
+      AsRobotUser asRobotUser = mock(AsRobotUser.class);
+      doCallRealMethod().when(asRobotUser).call(any());
       return asRobotUser;
     }
   }
